# Auto-vectorize  minmax instead of manual vectorization