--- a/stl/inc/__msvc_minmax.hpp
+++ b/stl/inc/__msvc_minmax.hpp
@@ -21,61 +21,6 @@ struct _Min_max_element_t {
     const void* _Min;
     const void* _Max;
 };
-
-struct _Min_max_1i {
-    int8_t _Min;
-    int8_t _Max;
-};
-
-struct _Min_max_1u {
-    uint8_t _Min;
-    uint8_t _Max;
-};
-
-struct _Min_max_2i {
-    int16_t _Min;
-    int16_t _Max;
-};
-
-struct _Min_max_2u {
-    uint16_t _Min;
-    uint16_t _Max;
-};
-
-struct _Min_max_4i {
-    int32_t _Min;
-    int32_t _Max;
-};
-
-struct _Min_max_4u {
-    uint32_t _Min;
-    uint32_t _Max;
-};
-
-struct _Min_max_8i {
-    int64_t _Min;
-    int64_t _Max;
-};
-
-struct _Min_max_8u {
-    uint64_t _Min;
-    uint64_t _Max;
-};
-
-struct _Min_max_f {
-    float _Min;
-    float _Max;
-};
-
-struct _Min_max_d {
-    double _Min;
-    double _Max;
-};
-
-struct _Min_max_p {
-    void* _Min;
-    void* _Max;
-};
 } // extern "C"
 
 #pragma pop_macro("new")