--- a/stl/inc/xutility
+++ b/stl/inc/xutility
@@ -121,27 +121,6 @@ const void* __stdcall __std_max_element_8(const void* _First, const void* _Last,
 const void* __stdcall __std_max_element_f(const void* _First, const void* _Last, bool _Unused) noexcept;
 const void* __stdcall __std_max_element_d(const void* _First, const void* _Last, bool _Unused) noexcept;
 
-__declspec(noalias) int8_t __stdcall __std_min_1i(const void* _First, const void* _Last) noexcept;
-__declspec(noalias) uint8_t __stdcall __std_min_1u(const void* _First, const void* _Last) noexcept;
-__declspec(noalias) int16_t __stdcall __std_min_2i(const void* _First, const void* _Last) noexcept;
-__declspec(noalias) uint16_t __stdcall __std_min_2u(const void* _First, const void* _Last) noexcept;
-__declspec(noalias) int32_t __stdcall __std_min_4i(const void* _First, const void* _Last) noexcept;
-__declspec(noalias) uint32_t __stdcall __std_min_4u(const void* _First, const void* _Last) noexcept;
-__declspec(noalias) int64_t __stdcall __std_min_8i(const void* _First, const void* _Last) noexcept;
-__declspec(noalias) uint64_t __stdcall __std_min_8u(const void* _First, const void* _Last) noexcept;
-__declspec(noalias) float __stdcall __std_min_f(const void* _First, const void* _Last) noexcept;
-__declspec(noalias) double __stdcall __std_min_d(const void* _First, const void* _Last) noexcept;
-__declspec(noalias) int8_t __stdcall __std_max_1i(const void* _First, const void* _Last) noexcept;
-__declspec(noalias) uint8_t __stdcall __std_max_1u(const void* _First, const void* _Last) noexcept;
-__declspec(noalias) int16_t __stdcall __std_max_2i(const void* _First, const void* _Last) noexcept;
-__declspec(noalias) uint16_t __stdcall __std_max_2u(const void* _First, const void* _Last) noexcept;
-__declspec(noalias) int32_t __stdcall __std_max_4i(const void* _First, const void* _Last) noexcept;
-__declspec(noalias) uint32_t __stdcall __std_max_4u(const void* _First, const void* _Last) noexcept;
-__declspec(noalias) int64_t __stdcall __std_max_8i(const void* _First, const void* _Last) noexcept;
-__declspec(noalias) uint64_t __stdcall __std_max_8u(const void* _First, const void* _Last) noexcept;
-__declspec(noalias) float __stdcall __std_max_f(const void* _First, const void* _Last) noexcept;
-__declspec(noalias) double __stdcall __std_max_d(const void* _First, const void* _Last) noexcept;
-
 __declspec(noalias) size_t __stdcall __std_mismatch_1(const void* _First1, const void* _First2, size_t _Count) noexcept;
 __declspec(noalias) size_t __stdcall __std_mismatch_2(const void* _First1, const void* _First2, size_t _Count) noexcept;
 __declspec(noalias) size_t __stdcall __std_mismatch_4(const void* _First1, const void* _First2, size_t _Count) noexcept;
@@ -327,49 +306,6 @@ auto _Min_vectorized(_Ty* const _First, _Ty* const _Last) noexcept {
             return ::__std_min_8i(_First, _Last);
         } else {
             return ::__std_min_8u(_First, _Last);
-        }
-    } else {
-        _STL_INTERNAL_STATIC_ASSERT(false); // unexpected size
-    }
-}
-
-template <class _Ty>
-auto _Max_vectorized(_Ty* const _First, _Ty* const _Last) noexcept {
-    constexpr bool _Signed = is_signed_v<_Ty>;
-
-    if constexpr (is_pointer_v<_Ty>) {
-#ifdef _WIN64
-        return reinterpret_cast<void*>(::__std_max_8u(_First, _Last));
-#else
-        return reinterpret_cast<void*>(::__std_max_4u(_First, _Last));
-#endif
-    } else if constexpr (is_same_v<remove_const_t<_Ty>, float>) {
-        return ::__std_max_f(_First, _Last);
-    } else if constexpr (_Is_any_of_v<remove_const_t<_Ty>, double, long double>) {
-        return ::__std_max_d(_First, _Last);
-    } else if constexpr (sizeof(_Ty) == 1) {
-        if constexpr (_Signed) {
-            return ::__std_max_1i(_First, _Last);
-        } else {
-            return ::__std_max_1u(_First, _Last);
-        }
-    } else if constexpr (sizeof(_Ty) == 2) {
-        if constexpr (_Signed) {
-            return ::__std_max_2i(_First, _Last);
-        } else {
-            return ::__std_max_2u(_First, _Last);
-        }
-    } else if constexpr (sizeof(_Ty) == 4) {
-        if constexpr (_Signed) {
-            return ::__std_max_4i(_First, _Last);
-        } else {
-            return ::__std_max_4u(_First, _Last);
-        }
-    } else if constexpr (sizeof(_Ty) == 8) {
-        if constexpr (_Signed) {
-            return ::__std_max_8i(_First, _Last);
-        } else {
-            return ::__std_max_8u(_First, _Last);
         }
     } else {
         _STL_INTERNAL_STATIC_ASSERT(false); // unexpected size
@@ -7030,20 +6966,31 @@ namespace ranges {
 #endif // _HAS_CXX20
 #endif // _HAS_CXX17
 
+template <class _Ty>
+constexpr _Ty _Max(const _Ty* _First, const _Ty* const _Last) {
+    // Vectorization-friendly maximum valiue
+    _Ty _Max_val = *_First;
+
+    for (++_First; _First != _Last; ++_First) {
+        if (_Max_val < *_First) {
+            _Max_val = *_First;
+        }
+    }
+
+    return _Max_val;
+}
+
 _EXPORT_STD template <class _Ty, class _Pr>
 _NODISCARD constexpr _Ty(max)(initializer_list<_Ty> _Ilist, _Pr _Pred) {
     // return leftmost/largest
     _STL_ASSERT(
         _Ilist.size() != 0, "An initializer_list passed to std::max must not be empty. (N4971 [alg.min.max]/13)");
-#if _USE_STD_VECTOR_ALGORITHMS
     if constexpr (_Is_min_max_value_optimization_safe<const _Ty*, _Pr>) {
-        if (!_Is_constant_evaluated()) {
-            return static_cast<_Ty>(_STD _Max_vectorized(_Ilist.begin(), _Ilist.end()));
-        }
+        return _STD _Max(_Ilist.begin(), _Ilist.end());
+    } else {
+        const _Ty* _Res = _STD _Max_element_unchecked(_Ilist.begin(), _Ilist.end(), _STD _Pass_fn(_Pred));
+        return *_Res;
     }
-#endif // _USE_STD_VECTOR_ALGORITHMS
-    const _Ty* _Res = _STD _Max_element_unchecked(_Ilist.begin(), _Ilist.end(), _STD _Pass_fn(_Pred));
-    return *_Res;
 }
 
 _EXPORT_STD template <class _Ty>
@@ -7081,14 +7028,11 @@ namespace ranges {
             const auto _Last  = _Range.end();
             _STL_ASSERT(_First != _Last,
                 "An initializer_list passed to std::ranges::max must not be empty. (N4971 [alg.min.max]/13)");
-#if _USE_STD_VECTOR_ALGORITHMS
             if constexpr (is_same_v<_Pj, identity> && _Is_min_max_value_optimization_safe<const _Ty*, _Pr>) {
-                if (!_STD is_constant_evaluated()) {
-                    return static_cast<_Ty>(_STD _Max_vectorized(_First, _Last));
-                }
+                return _STD _Max(_First, _Last);
+            } else {
+                return *_RANGES _Max_element_unchecked(_First, _Last, _STD _Pass_fn(_Pred), _STD _Pass_fn(_Proj));
             }
-#endif // _USE_STD_VECTOR_ALGORITHMS
-            return *_RANGES _Max_element_unchecked(_First, _Last, _STD _Pass_fn(_Pred), _STD _Pass_fn(_Proj));
         }
 
         template <input_range _Rng, class _Pj = identity,
@@ -7100,17 +7044,12 @@ namespace ranges {
             auto _ULast  = _Uend(_Range);
             _STL_ASSERT(
                 _UFirst != _ULast, "A range passed to std::ranges::max must not be empty. (N4971 [alg.min.max]/13)");
-#if _USE_STD_VECTOR_ALGORITHMS
             if constexpr (is_same_v<_Pj, identity> && _Is_min_max_value_optimization_safe<decltype(_UFirst), _Pr>
                           && sized_sentinel_for<decltype(_ULast), decltype(_UFirst)>) {
-                if (!_STD is_constant_evaluated()) {
-                    const auto _First_ptr = _STD to_address(_UFirst);
-                    const auto _Last_ptr  = _First_ptr + (_ULast - _UFirst);
-                    return static_cast<range_value_t<_Rng>>(_STD _Max_vectorized(_First_ptr, _Last_ptr));
-                }
-            }
-#endif // _USE_STD_VECTOR_ALGORITHMS
-            if constexpr (forward_range<_Rng> && _Prefer_iterator_copies<decltype(_UFirst)>) {
+                const auto _First_ptr = _STD to_address(_UFirst);
+                const auto _Last_ptr  = _First_ptr + (_ULast - _UFirst);
+                return static_cast<range_value_t<_Rng>>(_STD _Max(_First_ptr, _Last_ptr));
+            } else if constexpr (forward_range<_Rng> && _Prefer_iterator_copies<decltype(_UFirst)>) {
                 return static_cast<range_value_t<_Rng>>(*_RANGES _Max_element_unchecked(
                     _STD move(_UFirst), _STD move(_ULast), _STD _Pass_fn(_Pred), _STD _Pass_fn(_Proj)));
             } else {
@@ -7254,20 +7193,31 @@ namespace ranges {
 #endif // _HAS_CXX20
 #endif // _HAS_CXX17
 
+template <class _Ty>
+constexpr _Ty _Min(const _Ty* _First, const _Ty* const _Last) {
+    // Vectorization-friendly minimum valiue
+    _Ty _Min_val = *_First;
+
+    for (++_First; _First != _Last; ++_First) {
+        if (*_First < _Min_val) {
+            _Min_val = *_First;
+        }
+    }
+
+    return _Min_val;
+}
+
 _EXPORT_STD template <class _Ty, class _Pr>
 _NODISCARD constexpr _Ty(min)(initializer_list<_Ty> _Ilist, _Pr _Pred) {
     // return leftmost/smallest
     _STL_ASSERT(
         _Ilist.size() != 0, "An initializer_list passed to std::min must not be empty. (N4971 [alg.min.max]/5)");
-#if _USE_STD_VECTOR_ALGORITHMS
     if constexpr (_Is_min_max_value_optimization_safe<const _Ty*, _Pr>) {
-        if (!_Is_constant_evaluated()) {
-            return static_cast<_Ty>(_STD _Min_vectorized(_Ilist.begin(), _Ilist.end()));
-        }
+        return _STD _Min(_Ilist.begin(), _Ilist.end());
+    } else {
+        const _Ty* _Res = _STD _Min_element_unchecked(_Ilist.begin(), _Ilist.end(), _STD _Pass_fn(_Pred));
+        return *_Res;
     }
-#endif // _USE_STD_VECTOR_ALGORITHMS
-    const _Ty* _Res = _STD _Min_element_unchecked(_Ilist.begin(), _Ilist.end(), _STD _Pass_fn(_Pred));
-    return *_Res;
 }
 
 _EXPORT_STD template <class _Ty>
@@ -7299,14 +7249,11 @@ namespace ranges {
             const auto _Last  = _Range.end();
             _STL_ASSERT(_First != _Last,
                 "An initializer_list passed to std::ranges::min must not be empty. (N4971 [alg.min.max]/5)");
-#if _USE_STD_VECTOR_ALGORITHMS
             if constexpr (is_same_v<_Pj, identity> && _Is_min_max_value_optimization_safe<const _Ty*, _Pr>) {
-                if (!_STD is_constant_evaluated()) {
-                    return static_cast<_Ty>(_STD _Min_vectorized(_First, _Last));
-                }
+                return static_cast<_Ty>(_STD _Min(_First, _Last));
+            } else {
+                return *_RANGES _Min_element_unchecked(_First, _Last, _STD _Pass_fn(_Pred), _STD _Pass_fn(_Proj));
             }
-#endif // _USE_STD_VECTOR_ALGORITHMS
-            return *_RANGES _Min_element_unchecked(_First, _Last, _STD _Pass_fn(_Pred), _STD _Pass_fn(_Proj));
         }
 
         template <input_range _Rng, class _Pj = identity,
@@ -7318,17 +7265,12 @@ namespace ranges {
             auto _ULast  = _Uend(_Range);
             _STL_ASSERT(
                 _UFirst != _ULast, "A range passed to std::ranges::min must not be empty. (N4971 [alg.min.max]/5)");
-#if _USE_STD_VECTOR_ALGORITHMS
             if constexpr (is_same_v<_Pj, identity> && _Is_min_max_value_optimization_safe<decltype(_UFirst), _Pr>
                           && sized_sentinel_for<decltype(_ULast), decltype(_UFirst)>) {
-                if (!_STD is_constant_evaluated()) {
-                    const auto _First_ptr = _STD to_address(_UFirst);
-                    const auto _Last_ptr  = _First_ptr + (_ULast - _UFirst);
-                    return static_cast<range_value_t<_Rng>>(_STD _Min_vectorized(_First_ptr, _Last_ptr));
-                }
-            }
-#endif // _USE_STD_VECTOR_ALGORITHMS
-            if constexpr (forward_range<_Rng> && _Prefer_iterator_copies<decltype(_UFirst)>) {
+                const auto _First_ptr = _STD to_address(_UFirst);
+                const auto _Last_ptr  = _First_ptr + (_ULast - _UFirst);
+                return static_cast<range_value_t<_Rng>>(_STD _Min(_First_ptr, _Last_ptr));
+            } else if constexpr (forward_range<_Rng> && _Prefer_iterator_copies<decltype(_UFirst)>) {
                 return static_cast<range_value_t<_Rng>>(*_RANGES _Min_element_unchecked(
                     _STD move(_UFirst), _STD move(_ULast), _STD _Pass_fn(_Pred), _STD _Pass_fn(_Proj)));
             } else {