Closes #4453

# 🤔 Should this be done?

I'm not sure.

In #4453 there's a decision we should.

#4673 was closed without merging.

The benchmark results are ambiguous.

# 🏁 Benchmark results

All benchmarks were ran with `/fp:fast` to make the floating versions affected too. (This optimization is only engaged with `/fp;fast`)

Since auto-vectorization doesn't do runtime dispatch, it was separately ran with `/arch:AVX2` and with default options

Benchmark                           | main + /fp:fast | this + /fp:fast| this + /fp:fast+ AVX2
------------------------------------|-----------------|----------------|-------------------
bm<uint8_t, Op::Min_val>/8021       |  68.2 ns        |  93.5 ns       |   72.2 ns    
bm<uint8_t, Op::Min_val>/63         |  4.52 ns        |  15.3 ns       |   22.3 ns    
bm<uint8_t, Op::Max_val>/8021       |  77.9 ns        |  90.9 ns       |   70.9 ns    
bm<uint8_t, Op::Max_val>/63         |  4.41 ns        |  17.9 ns       |   22.3 ns    
bm<uint8_t, Op::Both_val>/8021      |  94.9 ns        |   143 ns       |    145 ns    
bm<uint8_t, Op::Both_val>/63        |  5.86 ns        |  21.2 ns       |   39.3 ns    
bm<uint16_t, Op::Min_val>/8021      |   129 ns        |   177 ns       |    135 ns    
bm<uint16_t, Op::Min_val>/31        |  3.61 ns        |  6.93 ns       |   12.5 ns    
bm<uint16_t, Op::Max_val>/8021      |   147 ns        |   157 ns       |    137 ns    
bm<uint16_t, Op::Max_val>/31        |  3.59 ns        |  12.6 ns       |   12.4 ns    
bm<uint16_t, Op::Both_val>/8021     |   166 ns        |   244 ns       |    276 ns    
bm<uint16_t, Op::Both_val>/31       |  4.53 ns        |  13.2 ns       |   24.3 ns    
bm<uint32_t, Op::Min_val>/8021      |   267 ns        |   310 ns       |    224 ns    
bm<uint32_t, Op::Min_val>/15        |  3.18 ns        |  4.23 ns       |   5.94 ns    
bm<uint32_t, Op::Max_val>/8021      |   243 ns        |   334 ns       |    238 ns    
bm<uint32_t, Op::Max_val>/15        |  3.13 ns        |  4.11 ns       |   7.77 ns    
bm<uint32_t, Op::Both_val>/8021     |   334 ns        |   489 ns       |    422 ns    
bm<uint32_t, Op::Both_val>/15       |  3.93 ns        |  6.25 ns       |   8.69 ns    
bm<uint64_t, Op::Min_val>/8021      |  2829 ns        |  3840 ns       |   3793 ns    
bm<uint64_t, Op::Min_val>/7         |  4.59 ns        |  3.10 ns       |   3.34 ns    
bm<uint64_t, Op::Max_val>/8021      |  2849 ns        |  3835 ns       |   3791 ns    
bm<uint64_t, Op::Max_val>/7         |  4.66 ns        |  3.10 ns       |   3.35 ns    
bm<uint64_t, Op::Both_val>/8021     |  3027 ns        |  5719 ns       |   5761 ns    
bm<uint64_t, Op::Both_val>/7        |  9.24 ns        |  4.50 ns       |   4.47 ns    
bm<int8_t, Op::Min_val>/8021        |  76.4 ns        |  90.2 ns       |   69.6 ns    
bm<int8_t, Op::Min_val>/63          |  4.63 ns        |  15.1 ns       |   20.1 ns    
bm<int8_t, Op::Max_val>/8021        |  69.0 ns        |   106 ns       |   72.1 ns    
bm<int8_t, Op::Max_val>/63          |  4.59 ns        |  11.2 ns       |   24.2 ns    
bm<int8_t, Op::Both_val>/8021       |  87.2 ns        |   141 ns       |    142 ns    
bm<int8_t, Op::Both_val>/63         |  7.11 ns        |  18.3 ns       |   34.6 ns    
bm<int16_t, Op::Min_val>/8021       |   124 ns        |   163 ns       |    137 ns    
bm<int16_t, Op::Min_val>/31         |  4.06 ns        |  7.83 ns       |   12.4 ns    
bm<int16_t, Op::Max_val>/8021       |   132 ns        |   154 ns       |    137 ns    
bm<int16_t, Op::Max_val>/31         |  4.07 ns        |  7.85 ns       |   12.4 ns    
bm<int16_t, Op::Both_val>/8021      |   176 ns        |   248 ns       |    271 ns    
bm<int16_t, Op::Both_val>/31        |  5.11 ns        |  12.6 ns       |   20.1 ns    
bm<int32_t, Op::Min_val>/8021       |   247 ns        |   332 ns       |    235 ns    
bm<int32_t, Op::Min_val>/15         |  3.62 ns        |  3.78 ns       |   6.83 ns    
bm<int32_t, Op::Max_val>/8021       |   263 ns        |   306 ns       |    233 ns    
bm<int32_t, Op::Max_val>/15         |  3.70 ns        |  3.97 ns       |   7.15 ns    
bm<int32_t, Op::Both_val>/8021      |   328 ns        |   486 ns       |    433 ns    
bm<int32_t, Op::Both_val>/15        |  4.18 ns        |  5.89 ns       |   8.47 ns    
bm<int64_t, Op::Min_val>/8021       |  2866 ns        |  3837 ns       |   3789 ns    
bm<int64_t, Op::Min_val>/7          |  3.83 ns        |  3.21 ns       |   3.46 ns    
bm<int64_t, Op::Max_val>/8021       |  2821 ns        |  3792 ns       |   3786 ns    
bm<int64_t, Op::Max_val>/7          |  3.83 ns        |  3.15 ns       |   3.42 ns    
bm<int64_t, Op::Both_val>/8021      |  3105 ns        |  3802 ns       |   3796 ns    
bm<int64_t, Op::Both_val>/7         |  8.94 ns        |  4.01 ns       |   4.27 ns    
bm<float, Op::Min_val>/8021         |   864 ns        |   906 ns       |    440 ns    
bm<float, Op::Min_val>/15           |  3.35 ns        |  4.39 ns       |   6.50 ns    
bm<float, Op::Max_val>/8021         |   872 ns        |   911 ns       |    446 ns    
bm<float, Op::Max_val>/15           |  3.34 ns        |  3.76 ns       |   4.27 ns    
bm<float, Op::Both_val>/8021        |  1062 ns        |   946 ns       |    476 ns    
bm<float, Op::Both_val>/15          |   111 ns        |  4.97 ns       |   6.49 ns    
bm<double, Op::Min_val>/8021        |  1819 ns        |  1872 ns       |    967 ns    
bm<double, Op::Min_val>/7           |  3.21 ns        |  2.60 ns       |   2.67 ns    
bm<double, Op::Max_val>/8021        |  1826 ns        |  1851 ns       |    939 ns    
bm<double, Op::Max_val>/7           |  3.24 ns        |  2.61 ns       |   2.66 ns    
bm<double, Op::Both_val>/8021       |  1997 ns        |  1890 ns       |   1166 ns    
bm<double, Op::Both_val>/7          |   113 ns        |  3.32 ns       |   3.43 ns    

# 🥈 Results observations

On large arrays for integer types manual vectorization of `min` and `max` has significant advantage for 64-bit, and some disadvantage for 32-bit. For `minmax`, manual vectorization is clearly better.

Small tails cases are way better for manual vectorization most of the time.

However, auto vectorization does surprisingly well on floats. 

Very odd anomaly is seen on `bm<float, Op::Both_val>/15` and `bm<double, Op::Both_val>/7` cases, I double checked that it is a consistent result.