@AlexGuteniev 2024-10-12
> However, auto vectorization does surprisingly well on floats.
> 
> Very odd anomaly is seen on `bm<float, Op::Both_val>/15` and `bm<double, Op::Both_val>/7` cases, I double checked that it is a consistent result.

### 🐞 First of all, this is a severe compiler bug that affects manually vectorized code. The compiler fails to place `vzeroupper` properly.

I'm afraid I don't have a workaround.

Also another thing smaller regarding codegen that I wasn't able to fix from my side.