@AlexGuteniev 2024-10-13
> ### 🐞 First of all, this is a severe compiler bug that affects manually vectorized code. The compiler fails to place `vzeroupper` properly.

Reported DevCom-10767462