@AlexGuteniev 2024-10-13
> Reported DevCom-10767462

This problems get unnoticed thus far because:
 * The original PR only had a benchmark for large array, where vectorization benefit outweighs the misplaced `vzeroupper` penalty
 * The recently merged masks PR added small arrays to benchmark masks, but it wasn't benchmarked with `/fp:fast` where floats+values case is executed
 * The test coverage doesn't catch this, as it is only perf problem that doesn't affect correcness