@AlexGuteniev 2024-10-13
> manual vectorization of `min` and `max` has significant advantage for 64-bit

This is obvious without even looing into it. The compiler doesn't attempt to vectorize 64-bit min/max. There are no 64-bit min/max instructions, and the compiler does not try `cmpgt` + `blendv`. 

The benefit from manual vectorization here  is small, less than x2, so maybe it doesn't worth a compiler optimization.