@AlexGuteniev 2024-10-13
Note that the whole problem with floating `minmax`, while being severe, seems rather rare:
 * It only affects `ranges::minmax`. The classic algorithms have only `_element` which is not affected.
 * It only affects `/fp:fast`. With other floating modes, including the default, floating `minmax` uses `_element` implementation to deal properly with negative zeros.