@AlexGuteniev 2024-10-15
> Also another thing smaller regarding codegen that I wasn't able to fix from my side.

Perhaps this would be fixable if we determine on keeping manual vectorization somehow; I've managed to generate proper insns sequence with intrinsics on Godbolt.