@StephanTLavavej 2024-10-16
We talked about this at the weekly maintainer meeting. Ignoring the enormous sunk cost of having reviewed all of the manual vectorization PRs for minmax, it would be generous to say that the performance results of auto-vectorization here are mixed - it sounds like manual vectorization is generally superior. As a result, this PR would be perceived as a performance regression. At this time, we believe that keeping the current approach is superior. Thanks for looking into this!