@AlexGuteniev 2024-11-03
I have looked into the cases where the auto vectorization wins. Unlike the manual vectorization, the auto one does the following things:
 * Loop unrolling by two iterations
 * Memory operand for min/max instruction instead of separate load and min/max

Looks like the speedup is contributed by entirely the first factor.

Not sure if manual vectorization should unroll loops.